Rename CollectionResource to ResourceCollection to better follow Laravel naming
