--- tests/Resources/AccountCollectionResource.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace Brainstud\JsonApi\Tests\Resources;
-
-use Brainstud\JsonApi\Resources\JsonApiCollectionResource;
-
-class AccountCollectionResource extends JsonApiCollectionResource
-{
-    public $collects = AccountResource::class;
-}
