+++ tests/Resources/AccountResourceCollection.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Brainstud\JsonApi\Tests\Resources;
+
+use Brainstud\JsonApi\Resources\JsonApiResourceCollection;
+
+class AccountResourceCollection extends JsonApiResourceCollection
+{
+    public $collects = AccountResource::class;
+}
