--- tests/Resources/CommentCollectionResource.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace Brainstud\JsonApi\Tests\Resources;
-
-use Brainstud\JsonApi\Resources\JsonApiCollectionResource;
-
-class CommentCollectionResource extends JsonApiCollectionResource
-{
-    public $collects = CommentResource::class;
-}
