--- tests/Resources/PostCollectionResource.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace Brainstud\JsonApi\Tests\Resources;
-
-use Brainstud\JsonApi\Resources\JsonApiCollectionResource;
-
-class PostCollectionResource extends JsonApiCollectionResource
-{
-    public $collects = PostResource::class;
-}
